ProducePricer/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.3.0
bleach==4.1.0
blinker==1.9.0
Brotli==1.2.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
contourpy==1.3.2
coverage==7.10.2
cryptography>=46.0.5
cssselect2==0.8.0
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Mailman==1.1.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.60.2
fpdf2==2.8.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
kiwisolver==1.4.8
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
numpy==2.2.5
openai==1.99.1
packaging==25.0
pandas==2.2.3
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.1.1
pluggy==1.6.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
Pygments==2.19.2
pyngrok==7.2.8
pyparsing==3.2.3
pypdfium2==4.30.0
pyphen==0.17.2
pytest==8.4.1
pytest-cov==6.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
qrcode>=8.0
RapidFuzz==3.13.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
tinycss2>=1.5.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
weasyprint==68.0
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.2.1
zopfli==0.2.3.post1